Move `ApiModules` interface to `module.ts`
